create shared_test and scope service tests
